+++ src/SqlArtisan/Internal/SqlPart/Expression/Function/AnalyticFunction/AnalyticPercentRankFunction.cs
@@ -4,12 +4,6 @@ public sealed class AnalyticPercentRankFunction : AnalyticFunction
 {
     internal AnalyticPercentRankFunction() { }
 
-    public WindowFunction Over(PartitionByAndOrderBy partitionByAndOrderBy) =>
-        new(this, OverClause.Of(partitionByAndOrderBy));
-
-    public WindowFunction Over(OrderByClause orderByClause) =>
-        new(this, OverClause.Of(orderByClause));
-
     internal override void Format(SqlBuildingBuffer buffer) => buffer
         .Append(Keywords.PercentRank)
         .OpenParenthesis()